+++ dlls/dll.cpp
@@ -1865,11 +1865,11 @@ void BotFunc_ReadProfile(std::FILE* fp, bot_profile_t* bpBotProfile)
 			bpBotProfile->m_OnosPercent = std::atoi(&szBuffer[13]);
 		}
 		else if (std::strncmp(szBuffer, "aim_speed=", 10) == 0)
-			bpBotProfile->m_fAimSpeed = std::atof(&szBuffer[10]);
+			bpBotProfile->m_fAimSpeed = static_cast<float>(std::atof(&szBuffer[10]));
 		else if (std::strncmp(szBuffer, "aim_skill=", 10) == 0)
-			bpBotProfile->m_fAimSkill = std::atof(&szBuffer[10]);
+			bpBotProfile->m_fAimSkill = static_cast<float>(std::atof(&szBuffer[10]));
 		else if (std::strncmp(szBuffer, "aim_time=", 9) == 0)
-			bpBotProfile->m_fAimTime = std::atof(&szBuffer[9]);
+			bpBotProfile->m_fAimTime = static_cast<float>(std::atof(&szBuffer[9]));
 		else if (std::strncmp(szBuffer, "bottomcolor=", 12) == 0)
 		{
 			bpBotProfile->m_iBottomColour = std::atoi(&szBuffer[12]);
@@ -1950,7 +1950,7 @@ void BotFunc_ReadProfile(std::FILE* fp, bot_profile_t* bpBotProfile)
 		}
 		else if (std::strncmp(szBuffer, "update_vision_time=", 19) == 0)
 		{
-			bpBotProfile->m_fVisionTime = std::atof(&szBuffer[19]);
+			bpBotProfile->m_fVisionTime = static_cast<float>(std::atof(&szBuffer[19]));
 		}
 	}
